agenticAI/requirements.txt
# Core dependencies
openai>=1.12.0
google-generativeai>=0.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# MCP Protocol support
mcp>=0.1.0

# Utilities
aiohttp>=3.9.0
tenacity>=8.2.0
typing-extensions>=4.9.0

# Development
pytest>=7.4.0
black>=23.12.0


